# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Spinlab Deal Desk API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /brokers:
    get:
      operationId: listBrokers
      tags: [brokers]
      summary: List all brokers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: market
          in: query
          schema:
            type: string
        - name: relationshipStrength
          in: query
          schema:
            type: string
        - name: followUpToday
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of brokers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Broker"
    post:
      operationId: createBroker
      tags: [brokers]
      summary: Create a new broker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrokerInput"
      responses:
        "201":
          description: Created broker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broker"
  /brokers/{id}:
    get:
      operationId: getBroker
      tags: [brokers]
      summary: Get a broker by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Broker details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broker"
        "404":
          description: Not found
    patch:
      operationId: updateBroker
      tags: [brokers]
      summary: Update a broker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrokerUpdate"
      responses:
        "200":
          description: Updated broker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broker"
    delete:
      operationId: deleteBroker
      tags: [brokers]
      summary: Delete a broker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /brokers/{id}/deals:
    get:
      operationId: getBrokerDeals
      tags: [brokers]
      summary: Get all deals linked to a broker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deals for broker
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"

  # ─── REMINDERS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Deal:
      type: object
      required:
        - id
        - dealName
        - status
        - priority
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        dealName:
          type: string
        businessName:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        assetType:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        brokerId:
          type: ["integer", "null"]
        brokerName:
          type: ["string", "null"]
        sellerName:
          type: ["string", "null"]
        askingPrice:
          type: ["number", "null"]
        grossRevenue:
          type: ["number", "null"]
        netIncome:
          type: ["number", "null"]
        adjustedNetIncome:
          type: ["number", "null"]
        monthlyRent:
          type: ["number", "null"]
        annualRent:
          type: ["number", "null"]
        leaseYearsRemaining:
          type: ["number", "null"]
        renewalOptions:
          type: ["string", "null"]
        squareFootage:
          type: ["number", "null"]
        rentPerSqFt:
          type: ["number", "null"]
        numWashers:
          type: ["integer", "null"]
        numDryers:
          type: ["integer", "null"]
        machineBrand:
          type: ["string", "null"]
        avgMachineAge:
          type: ["number", "null"]
        cardOrCoin:
          type: ["string", "null"]
        washAndFold:
          type: ["boolean", "null"]
        pickupDelivery:
          type: ["boolean", "null"]
        commercialAccounts:
          type: ["boolean", "null"]
        hoursOfOperation:
          type: ["string", "null"]
        staffCount:
          type: ["integer", "null"]
        ownerOperated:
          type: ["boolean", "null"]
        status:
          type: string
        priority:
          type: string
        nextAction:
          type: ["string", "null"]
        nextActionDueDate:
          type: ["string", "null"]
        lastContactedDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        askingMultiple:
          type: ["number", "null"]
        rentAsPercentGross:
          type: ["number", "null"]
        redFlagScore:
          type: integer
        redFlagLevel:
          type: string
        dealScore:
          type: integer
        dealQuality:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Broker:
      type: object
      required:
        - id
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        market:
          type: ["string", "null"]
        specialty:
          type: ["string", "null"]
        brokerType:
          type: ["string", "null"]
        relationshipStrength:
          type: ["string", "null"]
        trustRating:
          type: ["integer", "null"]
        responsivenessRating:
          type: ["integer", "null"]
        dealQualityRating:
          type: ["integer", "null"]
        firstContactedDate:
          type: ["string", "null"]
        lastContactedDate:
          type: ["string", "null"]
        nextFollowUpDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        dealCount:
          type: integer
        hotDealCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    BrokerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        company:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        market:
          type: string
        specialty:
          type: string
        brokerType:
          type: string
        relationshipStrength:
          type: string
        trustRating:
          type: ["integer", "null"]
        responsivenessRating:
          type: ["integer", "null"]
        dealQualityRating:
          type: ["integer", "null"]
        firstContactedDate:
          type: ["string", "null"]
        lastContactedDate:
          type: ["string", "null"]
        nextFollowUpDate:
          type: ["string", "null"]
        notes:
          type: string
        tags:
          type: string
    BrokerUpdate:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        market:
          type: string
        specialty:
          type: string
        brokerType:
          type: string
        relationshipStrength:
          type: string
        trustRating:
          type: ["integer", "null"]
        responsivenessRating:
          type: ["integer", "null"]
        dealQualityRating:
          type: ["integer", "null"]
        firstContactedDate:
          type: ["string", "null"]
        lastContactedDate:
          type: ["string", "null"]
        nextFollowUpDate:
          type: ["string", "null"]
        notes:
          type: string
        tags:
          type: string

    # ─── REMINDER ──────────────────────────────────────────────────────────
...