# lib/api-zod/src/generated/types/broker.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spinlab Deal Desk API
 * OpenAPI spec version: 0.1.0
 */

export interface Broker {
  id: number;
  name: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  market?: string | null;
  /** @nullable */
  specialty?: string | null;
  /** @nullable */
  brokerType?: string | null;
  /** @nullable */
  relationshipStrength?: string | null;
  /** @nullable */
  trustRating?: number | null;
  /** @nullable */
  responsivenessRating?: number | null;
  /** @nullable */
  dealQualityRating?: number | null;
  /** @nullable */
  firstContactedDate?: string | null;
  /** @nullable */
  lastContactedDate?: string | null;
  /** @nullable */
  nextFollowUpDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tags?: string | null;
  dealCount?: number;
  hotDealCount?: number;
  createdAt: string;
  updatedAt: string;
}