# lib/api-zod/src/generated/types/brokerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spinlab Deal Desk API
 * OpenAPI spec version: 0.1.0
 */

export interface BrokerInput {
  name: string;
  company?: string;
  phone?: string;
  email?: string;
  website?: string;
  market?: string;
  specialty?: string;
  brokerType?: string;
  relationshipStrength?: string;
  /** @nullable */
  trustRating?: number | null;
  /** @nullable */
  responsivenessRating?: number | null;
  /** @nullable */
  dealQualityRating?: number | null;
  /** @nullable */
  firstContactedDate?: string | null;
  /** @nullable */
  lastContactedDate?: string | null;
  /** @nullable */
  nextFollowUpDate?: string | null;
  notes?: string;
  tags?: string;
}