# lib/api-zod/src/generated/types/deal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Spinlab Deal Desk API
 * OpenAPI spec version: 0.1.0
 */

export interface Deal {
  id: number;
  dealName: string;
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  assetType?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  brokerId?: number | null;
  /** @nullable */
  brokerName?: string | null;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  askingPrice?: number | null;
  /** @nullable */
  grossRevenue?: number | null;
  /** @nullable */
  netIncome?: number | null;
  /** @nullable */
  adjustedNetIncome?: number | null;
  /** @nullable */
  monthlyRent?: number | null;
  /** @nullable */
  annualRent?: number | null;
  /** @nullable */
  leaseYearsRemaining?: number | null;
  /** @nullable */
  renewalOptions?: string | null;
  /** @nullable */
  squareFootage?: number | null;
  /** @nullable */
  rentPerSqFt?: number | null;
  /** @nullable */
  numWashers?: number | null;
  /** @nullable */
  numDryers?: number | null;
  /** @nullable */
  machineBrand?: string | null;
  /** @nullable */
  avgMachineAge?: number | null;
  /** @nullable */
  cardOrCoin?: string | null;
  /** @nullable */
  washAndFold?: boolean | null;
  /** @nullable */
  pickupDelivery?: boolean | null;
  /** @nullable */
  commercialAccounts?: boolean | null;
  /** @nullable */
  hoursOfOperation?: string | null;
  /** @nullable */
  staffCount?: number | null;
  /** @nullable */
  ownerOperated?: boolean | null;
  status: string;
  priority: string;
  /** @nullable */
  nextAction?: string | null;
  /** @nullable */
  nextActionDueDate?: string | null;
  /** @nullable */
  lastContactedDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  askingMultiple?: number | null;
  /** @nullable */
  rentAsPercentGross?: number | null;
  redFlagScore?: number;
  redFlagLevel?: string;
  dealScore?: number;
  dealQuality?: string;
  createdAt: string;
  updatedAt: string;
}